Data_Engineering/requirements.txt
# Required packages for Data Engineering Pipeline - Core Components Only

# Document processing
python-docx==1.1.0
docx2python==2.7.3
PyMuPDF==1.23.26
pdfplumber==0.10.3
openpyxl

# Text processing
chardet==5.2.0
ftfy==6.2.0
regex>=2022.1.18

# NLP - Core components
spacy==3.7.2
stanza==1.8.1
fr_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_lg-3.8.0/fr_core_news_lg-3.8.0-py3-none-any.whl
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
scikit-learn>=1.3.0

# Memory monitoring
psutil>=5.9.4

# Data handling
pandas
numpy==1.24.3  # Specific version to avoid compatibility issues with spaCy
jsonschema>=4.0.0

# Testing
pytest==8.0.0
